Moved states from buoy AI
